requirements.txt
-i https://pypi.org/simple
cramjam==2.6.2 ; python_version >= '3.7'
fastparquet==2022.11.0
fsspec==2022.11.0 ; python_version >= '3.7'
numpy==1.23.5 ; python_version >= '3.8'
packaging==21.3 ; python_version >= '3.6'
pandas==1.5.2
pyarrow==10.0.1
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.6
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytorch-lightning==1.7.4
scipy==1.10.1
matplotlib==3.7.1
scikit-learn==1.2.2
optuna==3.1.1
statsmodels==0.13.5
tensorboard==2.12.2